+++ src/Constants.ts
@@ -1,6 +1,7 @@
 export interface Options {
     DBName?: string;
     DBPath?: string;
+    checkVersion?: boolean;
     rewards: Rewards;
 }